5月26日晚提交，修改了很多button的显示

+++ ChuangYeBa/ArticleInfo.h
@@ -26,6 +26,6 @@
 @property (copy, nonatomic) NSNumber *collects;
 @property (strong, nonatomic) NSDate *publishDate;
 
-- (float)getHeightOfArticleString:(NSString *)string lineSpacing:(NSUInteger)lineSpacing fontOfSize:(NSUInteger)fontOfSize widthOffset:(float)widthOffset;
+- (NSInteger)getHeightOfArticleString:(NSString *)string lineSpacing:(NSUInteger)lineSpacing fontOfSize:(NSUInteger)fontOfSize widthOffset:(float)widthOffset;
 
 @end
