+++ ChuangYeBa/ArticleInfo.m
@@ -7,6 +7,7 @@
 //
 
 #import "ArticleInfo.h"
+#import <math.h>
 
 @implementation ArticleInfo
 
@@ -51,7 +52,7 @@
 }
 
 
-- (float)getHeightOfArticleString:(NSString *)string lineSpacing:(NSUInteger)lineSpacing fontOfSize:(NSUInteger)fontOfSize widthOffset:(float)widthOffset {
+- (NSInteger)getHeightOfArticleString:(NSString *)string lineSpacing:(NSUInteger)lineSpacing fontOfSize:(NSUInteger)fontOfSize widthOffset:(float)widthOffset {
     // 计算问题文本的高度
     // TODO 应该加上添加margin的输入参数
     NSMutableParagraphStyle *paragraphStyle = [[NSMutableParagraphStyle alloc] init];
@@ -62,7 +63,7 @@
     // 加上文本高度
     float height = frame.size.height;
     // 文本高度应该加上一点点边？？
-    return height + 24;
+    return (NSInteger)height + 24;
 }
 
 @end
